abstract chat logic
