+++ nucliadb/nucliadb/search/api/v1/chat.py
@@ -17,31 +17,20 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
-import base64
-from typing import AsyncIterator, List, Union
+from typing import Union
 
 from fastapi import Body, Header, Request, Response
 from fastapi_versioning import version
-from nucliadb_protos.nodereader_pb2 import RelationSearchRequest, RelationSearchResponse
 from starlette.responses import StreamingResponse
 
 from nucliadb.models.responses import HTTPClientError
 from nucliadb.search.api.v1.find import find
 from nucliadb.search.api.v1.router import KB_PREFIX, api
-from nucliadb.search.predict import PredictEngine
-from nucliadb.search.requesters.utils import Method, node_query
-from nucliadb.search.search.chat.prompt import format_chat_prompt_content
-from nucliadb.search.search.merge import merge_relations_results
-from nucliadb.search.utilities import get_predict
+from nucliadb.search.search.chat.query import chat, rephrase_query_from_context
 from nucliadb_models.resource import NucliaDBRoles
 from nucliadb_models.search import (
-    Author,
-    ChatModel,
-    ChatOptions,
     ChatRequest,
     FindRequest,
-    KnowledgeboxFindResults,
-    Message,
     NucliaDBClientType,
     SearchOptions,
 )
@@ -85,7 +74,7 @@ async def chat_post_knowledgebox(
     x_forwarded_for: str = Header(""),
 ) -> Union[StreamingResponse, HTTPClientError]:
     try:
-        return await chat(
+        return await chat_on_knowledgebox(
             response, kbid, item, x_ndb_client, x_nucliadb_user, x_forwarded_for
         )
     except LimitsExceededError as exc:
@@ -97,7 +86,7 @@ async def chat_post_knowledgebox(
         )
 
 
-async def chat(
+async def chat_on_knowledgebox(
     response: Response,
     kbid: str,
     item: ChatRequest,
@@ -105,22 +94,13 @@ async def chat(
     x_nucliadb_user: str,
     x_forwarded_for: str,
 ):
-    predict = get_predict()
-
     if item.context is not None and len(item.context) > 0:
-        # There is context lets do a query
-        req = ChatModel(
-            question=item.query,
-            context=item.context,
-            user_id=x_nucliadb_user,
-            retrieval=False,
+        new_query = await rephrase_query_from_context(
+            kbid, item.context, item.query, x_nucliadb_user
         )
-
-        new_query = await predict.rephrase_query(kbid, req)
     else:
         new_query = item.query
 
-    # ONLY PARAGRAPHS I VECTORS
     find_request = FindRequest()
     find_request.features = [
         SearchOptions.PARAGRAPH,
@@ -141,7 +121,7 @@ async def chat(
     find_request.autofilter = item.autofilter
     find_request.highlight = item.highlight
 
-    results, incomplete = await find(
+    find_results, incomplete = await find(
         response,
         kbid,
         find_request,
@@ -151,76 +131,4 @@ async def chat(
     )
     if incomplete:
         raise IncompleteFindResultsError()
-
-    if item.context is None:
-        context = []
-    else:
-        context = item.context
-    context.append(
-        Message(
-            author=Author.NUCLIA, text=await format_chat_prompt_content(kbid, results)
-        )
-    )
-
-    chat_model = ChatModel(
-        user_id=x_nucliadb_user,
-        context=context,
-        question=item.query,
-    )
-
-    ident, generator = await predict.chat_query(kbid, chat_model)
-
-    async def generate_answer(
-        results: KnowledgeboxFindResults,
-        kbid: str,
-        predict: PredictEngine,
-        generator: AsyncIterator[bytes],
-        features: List[ChatOptions],
-    ):
-        if ChatOptions.PARAGRAPHS in features:
-            bytes_results = base64.b64encode(results.json().encode())
-            yield len(bytes_results).to_bytes(length=4, byteorder="big", signed=False)
-            yield bytes_results
-
-        answer = []
-        async for data in generator:
-            answer.append(data)
-            yield data
-
-        if ChatOptions.RELATIONS in features:
-            yield END_OF_STREAM
-
-            text_answer = b"".join(answer)
-
-            detected_entities = await predict.detect_entities(
-                kbid, text_answer.decode()
-            )
-            relation_request = RelationSearchRequest()
-            relation_request.subgraph.entry_points.extend(detected_entities)
-            relation_request.subgraph.depth = 1
-
-            relations_results: List[RelationSearchResponse]
-            (
-                relations_results,
-                incomplete_results,
-                queried_nodes,
-                queried_shards,
-            ) = await node_query(kbid, Method.RELATIONS, relation_request, item.shards)
-            yield base64.b64encode(
-                (
-                    await merge_relations_results(
-                        relations_results, relation_request.subgraph
-                    )
-                )
-                .json()
-                .encode()
-            )
-
-    return StreamingResponse(
-        generate_answer(results, kbid, predict, generator, item.features),
-        media_type="plain/text",
-        headers={
-            "NUCLIA-LEARNING-ID": ident or "unknown",
-            "Access-Control-Expose-Headers": "NUCLIA-LEARNING-ID",
-        },
-    )
+    return await chat(kbid, find_results, item, x_nucliadb_user)
